requirements.txt
streamlit
tensorflow
opencv-python
streamlit-webrtc
pillow
numpy
mediapipe